requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
mcp>=1.0.0
GitPython>=3.1.40
pathspec>=0.12.1
